% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FullCTP_C}
\alias{FullCTP_C}
\title{Leading NA}
\usage{
FullCTP_C(LocalTest, f, pvals, EarlyStop, alpha)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{f}{A function that iterates LocalTest over the relevant test tree.
In practice, this is called as TestSet_C.}

\item{pvals}{A vector of p-values.}

\item{EarlyStop}{Logical indicating whether to exit as soon as a non-significant
p-value is found.}

\item{alpha}{Significance level. This is only used if EarlyStop = TRUE}
}
\description{
Tests a user-specified subset in a CTP, using a user-supplied local test
}
