% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FWER_set_C}
\alias{FWER_set_C}
\title{Leading NA}
\usage{
FWER_set_C(LocalTest, pvals, alpha, low, high, verbose)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{pvals}{A vector of p-values.}

\item{alpha}{A double indicating the significance level}

\item{low}{integer denoting the starting point for the search. Should start at zero.}

\item{high}{integer denoting the end point of the search. Should end at pvals.size() - 1.}

\item{verbose}{boolean, indicating whether to print progress.}
}
\value{
The number of hypotheses that can be rejected with kFWER control at a user
specific k.
}
\description{
Computes a the number of hypotheses that can be rejected with FWER control by
using a binary search
}
