% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kFWER_set_C}
\alias{kFWER_set_C}
\title{Leading NA}
\usage{
kFWER_set_C(LocalTest, pvals, k, alpha, low, high, verbose)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{pvals}{A vector of p-values.}

\item{k}{integer denoting the k to control the kFWER at.}

\item{alpha}{A double indicating the significance level}

\item{low}{integer denoting the starting point for the search. Should start at zero.}

\item{high}{integer denoting the end point of the search. Should end at pvals.size() - 1.}

\item{verbose}{boolean, indicating whether to print progress.}
}
\value{
The number of hypotheses that can be rejected with kFWER control at a user
specific k.
}
\description{
Computes a confidence set for the number of false hypotheses among a subset of
using a binary search
}
