% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MakeY_C}
\alias{MakeY_C}
\title{Leading NA}
\usage{
MakeY_C(pvals, m)
}
\arguments{
\item{pvals}{A NumericVector containing the truncated sorted p-values. It
is important that this vector: 1) contains only the truncated p-values (i.e,
those that fall below the truncation point) and 2) is sorted.}

\item{m}{The total (i.e., non-truncated) number of p-values.}
}
\description{
Returns the transformed p-values (Y) from pre-sorted p-values and
pre-truncated p-values. If not truncation is used, set m_full = m
}
