% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MakeZ_C}
\alias{MakeZ_C}
\title{Leading NA}
\usage{
MakeZ_C(pvals, m)
}
\arguments{
\item{pvals}{A NumericVector containing the truncated sorted p-values. It
is important that this vector: 1) contains only the truncated p-values (i.e,
those that fall below the truncation point) and 2) is sorted.}

\item{m}{The total (i.e., non-truncated) number of p-values.}
}
\description{
Returns the TMTI_infinity statistic from pre-sorted,
pre-truncated vector of p-values. If no truncation is used, set m_full = m
}
