% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MakeZ_C_nsmall}
\alias{MakeZ_C_nsmall}
\title{Leading NA}
\usage{
MakeZ_C_nsmall(pvals, n, m)
}
\arguments{
\item{pvals}{A NumericVector containing the truncated sorted p-values. It
is important that this vector: 1) contains only the truncated p-values (i.e,
those that fall below the truncation point) and 2) is sorted.}

\item{n}{A positive number (or Inf) indicating which type of local minimum
to consider. Defaults to Infm, corresponding to the global minimum.}

\item{m}{The total (i.e., non-truncated) number of p-values.}
}
\description{
Returns the transformed p-values (Y) from pre-sorted p-values and
pre-truncated p-values when n < m - 1
}
