% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtTMTI_CDF.R
\name{rtTMTI_CDF}
\alias{rtTMTI_CDF}
\title{Computes the analytical version of the rtMTI_infty CDF. When m>100, this should
not be used.}
\usage{
rtTMTI_CDF(x, m, K)
}
\arguments{
\item{x}{Point in which to evaluate the CDF}

\item{m}{Number of independent tests to combine}

\item{K}{Integer; the truncation point to use.}
}
\value{
The probability that the test statistic is at most x assuming
independence under the global null hypothesis.
}
\description{
Computes the analytical version of the rtMTI_infty CDF. When m>100, this should
not be used.
}
\examples{
rtTMTI_CDF(0.05, 100, 10)
}
