% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestSet_C}
\alias{TestSet_C}
\title{Leading NA}
\usage{
TestSet_C(
  LocalTest,
  pSub,
  pRest,
  alpha,
  is_subset_sequence,
  EarlyStop,
  verbose
)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{pSub}{A vector with the p-values of the set to be tested.}

\item{pRest}{A vector containing the remaining p-values.}

\item{alpha}{Double indicating the significance level.}

\item{is_subset_sequence}{Logical indicating whether the supplied subset of
p_values corresponds to the pSub.size() smallest overall p-values.}

\item{EarlyStop}{Logical indicating whether to exit as soon as a non-significant
p-value is found.}

\item{verbose}{Logical indicating whether to print progress.}
}
\description{
Tests a user-specified subset in a CTP, using a user-supplied local test
}
