% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMTI_CDF.R
\name{TMTI_CDF}
\alias{TMTI_CDF}
\title{Computes the analytical version of the TMTI_infty CDF. When m>100, this should
not be used.}
\usage{
TMTI_CDF(x, m)
}
\arguments{
\item{x}{Point in which to evaluate the CDF.}

\item{m}{Number of independent tests to combine.}
}
\value{
The probability that the test statistic is at most x assuming
independence under the global null hypothesis.
}
\description{
Computes the analytical version of the TMTI_infty CDF. When m>100, this should
not be used.
}
\examples{
TMTI_CDF(0.05, 100)
}
