% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TopDown_C_binary_subset}
\alias{TopDown_C_binary_subset}
\title{Leading NA}
\usage{
TopDown_C_binary_subset(LocalTest, pSub, pRest, alpha, low, high, verbose)
}
\arguments{
\item{LocalTest}{A function that returns a double in (0, 1).}

\item{pSub}{A vector of p-values from the subset of interest.}

\item{pRest}{A vector of the remaining p-values.}

\item{alpha}{A double indicating the significance level}

\item{low}{integer denoting the starting point for the search. Should start at zero.}

\item{high}{integer denoting the end point of the search. Should end at pvals.size() - 1.}

\item{verbose}{boolean, indicating whether to print progress.}
}
\description{
Computes a confidence set for the number of false hypotheses among a subset of
using a binary search
}
